requirements.txt
anyio==3.7.1
async-lru==1.0.2
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
dnspython==2.4.0
exceptiongroup==1.1.2
Flask==2.3.2
h11==0.14.0
httpcore==0.17.3
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
pyaes==1.6.1
pymongo==4.4.1
Pyrogram==2.0.106
PySocks==1.7.1
python-dotenv==1.0.0
requests==2.31.0
sniffio==1.3.0
TgCrypto==1.2.5
urllib3==2.0.4
Werkzeug==2.3.6
